pyproject.toml
[project]
name = "autoaiapply"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "chromium>=0.0.0",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "transformers>=4.47.1",
    "windows-curses>=2.4.0",
]
